Change the generated commands script.
Commands and their type info are now kept in a hash of function names to
hashes of data types (GL_COMMAND_TYPES). This effectively allows me to query a function's info rather than just loading it and hoping it exists. As a result, the loading code can be reduce to a call to __load_gl_sym__, which is also included in the generated code. I should probably break it off into the GlSym code but for now this'll do.

The upshot to this is that loading a symbol no longer raises exceptions
if it doesn't exist. It simply returns the wrapped function if loaded or
false if not. So, if you load a symbol and it's not there, the exception
only comes from attempting to use that symbol. So, it's better now to
query whether the function is loaded first rather than using it if I
can't be certain it's available.

Also, Gl#have_gl_command? is now a thing, so the above query is now doable as well. There's also Gl#load_all_gl_commands!, which does what it says on the tin.

--

GlSym is now a submodule of Gl. __load_gl_sym__ caches the loaded function. It also stores the command types because I didn't want to make those part of the Gl module (and thereby accidentally infect the global namespace if someone does `include Gl`, which sounds pretty reasonable to me).