Note about thread-safety of have_gl_command? and load_all_gl_commands!
Not really necessary, given that this is loading commands at runtime,
but Ruby folks may not be as conscious of what counts as thread-safe and
so on. So, here's the deal: if a Gl command is being loaded, one cannot
call have_gl_command? or otherwise call an unloaded function from
another thread. This is because there is no mutex around the Gl function
Hash and so on.

If you want some amount of thread-safety, call load_all_gl_commands!
before any threads are spawned. This will ensure that no Gl commands are
loaded in the future unless you do something weird like modify
GL_COMMAND_TYPES and add a new command, which _you should not do_.