Refactor the command/enum removal stuff a bit.
This removes the overly long and obtuse xpath strings, meaning there are
some that're not in constants now, but it's a lot more specific as well.

Mostly did this to make sure GL 4.4 was working right, since some
commands disappeared from gl_commands.rb, but it looks like they maybe
shouldn't have been there in the first place.