Fix uniform location caching issue with hint_uniform.
hint_uniform set a uniform's cached location to -1, causing it to fail
to reload. This is mainly because __reload_uniforms__ used
uniform_location to get a new uniform location, but that doesn't really
work out because uniform_location only returns cached values (or gets a
uniform's location, if not already cached, and caches it).